Revert "Merge branch 'main' of https://github.com/RongHua778/DBGTowerDefense into main"
This reverts commit d5edcbbedaec1078b21d5b99fa1a68a5c2976f55, reversing
changes made to 20630ef9354b283ec5cc13e69ff16c6ceb03a582.

+++ NEWTEST/Assets/Scripts/ScriptableObject/CardSO.cs
@@ -82,7 +82,7 @@ public class CardSO : ScriptableObject
     [Header("TowerCard Info")]
     public GameObject TurretPrefab;
     public float Damage;
-    public int Range;
+    public float Range;
     public float Speed;
     public float PersistTime;
     public float CriticalRate;
